Split controls into separate files
